+++ artemis/src/main/kotlin/ktx/artemis/arrays.kt
@@ -1,0 +1,6 @@
+package ktx.artemis
+
+import kotlin.reflect.KClass
+
+internal fun <T : Any> Array<out KClass<out T>>.asJavaClasses(): Array<Class<out T>> =
+  Array(size) { index -> this[index].java }