+++ artemis/src/test/kotlin/ktx/artemis/Dead.kt
@@ -1,0 +1,5 @@
+package ktx.artemis
+
+import com.artemis.Component
+
+class Dead : Component()